requirements.txt
absl-py==0.8.1
astor==0.8.0
beautifulsoup4==4.6.0
blis==0.2.4
boto==2.49.0
boto3==1.10.39
botocore==1.13.39
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
Click==7.0
colorama==0.4.1
cryptography==2.8
cymem==2.0.2
docutils==0.15.2
Flask==1.1.1
gast==0.3.2
gensim==3.8.1
greenlet==0.4.15
grpcio==1.25.0
hanja==0.13.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
JPype1==0.7.0
konlpy==0.5.2
korean==0.1.8
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
mathparse==0.1.2
msgpack==0.6.1
murmurhash==1.0.2
nltk==3.4.5
numpy==1.14.0
oauthlib==3.1.0
pandas==0.25.3
pbr==5.4.4
Pillow==6.2.1
Pint==0.9
plac==0.9.6
preshed==2.0.1
protobuf==3.11.1
pycparser==2.19
pymongo==3.9.0
pynvim==0.3.2
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.7.5
pytz==2019.3
PyYAML==3.13
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
sklearn==0.0
smart-open==1.9.0
spacy==2.1.8
SQLAlchemy==1.2.19
srsly==0.1.0
stevedore==1.31.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
thinc==7.0.8
torch==1.2.0
tqdm==4.36.1
tweepy==3.8.0
urllib3==1.25.5
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wasabi==0.2.2
Werkzeug==0.16.0
